Refactor rspec by isolating OpenTelemetry context propagation logic
